schema: Add migration `0002_schema.up.sql`
`0002_schema.up.sql` changes the PRIMARY KEY of `block`
table from `(height, hash)` to `height` only, in order
to prevent multiple rows from contending for the same height.

**NOTE** this operation is potentially destructive. If two
or more rows existed at the same height, they'll now be
missing from the database since we can't reliably determine
which block hash is the correct one.

Additionally, the anchor `block_hash` column previously
had the type INTEGER, and this was a mistake in the schema,
since block hashes are typically represented as a hex string.
Thus we change the type of the `block_hash` column to TEXT.